refactor(#792): split project_admin_persistence_tests into base + NAM/command suite (sweep)
project_admin_persistence_tests.rs 1023 -> 568, + project_admin_nam_tests.rs 469
(both under cap; off debt). The NAM-install / command-driven admin tests move to the
sibling; the Sandbox fixture (struct + new/new_session/save/reload + cfg field) and
the chain helpers become pub(super), imported via super. adapter-gui lib tests (540)
pass, zero warnings.
